separate OpenGL program and shader implementation
