+++ src/client/drawing/gl_es/primitive.cpp
@@ -1,5 +1,5 @@
 #include "primitive.hpp"
-#include "../gl_shared/shader_and_program.hpp"
+#include "../gl_shared/program.hpp"
 
 namespace Drawing
 {
