+++ src/client/drawing/gl_shared/program.cpp
@@ -1,7 +1,8 @@
-#include "shader_and_program.hpp"
-#include <SDL.h>
+#include "program.hpp"
+#include <SDL_log.h>
 
 #include "common.hpp"
+#include "shader.hpp"
 
 namespace Drawing
 {
@@ -11,46 +12,6 @@ namespace Detail
 
 namespace GLShared
 {
-
-// Shader IMPLEMENTATION
-
-Shader::Shader(const GLenum type, std::string_view source)
-	: type(type)
-{
-	ref = glCreateShader(type);
-	if(ref == 0)
-		return; // TODO: add log info
-	const char* src[] = {source.data()};
-	glShaderSource(ref, 1, src, NULL);
-	glCompileShader(ref);
-	GLint success;
-	glGetShaderiv(ref, GL_COMPILE_STATUS, &success);
-	if(success == GL_FALSE)
-	{
-		GLint logLength;
-		glGetShaderiv(ref, GL_INFO_LOG_LENGTH, &logLength);
-		auto logText = new char[logLength];
-		glGetShaderInfoLog(ref, logLength, NULL, logText);
-		SDL_LogError(SDL_LOG_CATEGORY_APPLICATION,
-		             "Could not compile shader: %s", logText);
-		delete[] logText;
-		return;
-	}
-	compiled = true;
-}
-
-Shader::~Shader()
-{
-	// NOTE: From reference: A value of 0 for shader will be silently ignored.
-	glDeleteShader(ref);
-}
-
-GLuint Shader::GetGLRef() const
-{
-	return ref;
-}
-
-// Program IMPLEMENTATION
 
 Program::Program()
 {
